requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.5.1+cu121
trl[vllm]==0.16.0
transformers==4.50.3
vllm==0.7.3

hydra-core
omegaconf
rich
wandb
